pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "png2pptx"
version = "0.1.0"
description = "Convert infographic-style PNGs into editable PowerPoint slides"
readme = { file = "README.md", content-type = "text/markdown" }
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "png2pptx contributors" }]
maintainers = [{ name = "png2pptx contributors" }]
requires-python = ">=3.10"
keywords = ["png", "pptx", "powerpoint", "ocr", "tesseract", "infographic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0",
    "numpy>=1.24,<2",
    "opencv-python>=4.5",
    "python-pptx>=0.6.21",
    "click>=8.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "build>=1.2"]

[project.scripts]
png2pptx = "png2pptx.cli:main"

[tool.setuptools.packages.find]
include = ["png2pptx*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
